extracting portfolio projects to a separate file
