+++ src/components/Projects/index.jsx
@@ -11,6 +11,7 @@ import Typography from "@material-ui/core/Typography";
 import VisibilityOutlinedIcon from "@material-ui/icons/VisibilityOutlined";
 import Palm from "../../images/palms.jpg";
 import routes from "../../routes/routes";
+import portfolio from "../../portfolio";
 
 const useStyles = makeStyles(theme => ({
   root: {
@@ -46,24 +47,6 @@ const useStyles = makeStyles(theme => ({
   }
 }));
 
-const portfolio = [
-  {
-    project_id: "1",
-    title: "Project 1",
-    desc: "This project is regarding payments"
-  },
-  {
-    project_id: "2",
-    title: "Project 2",
-    desc: "This project is regarding payments"
-  },
-  {
-    project_id: "3",
-    title: "Project 3",
-    desc: "This project is regarding payments"
-  }
-];
-
 function Projects() {
   const classes = useStyles();
 
@@ -83,7 +66,7 @@ function Projects() {
                 <CardActionArea>
                   <CardMedia
                     className={classes.media}
-                    image={Palm}
+                    image={project.image}
                     title={project.title}
                   />
                   <CardContent>
