+++ src/portfolio.js
@@ -1,0 +1,22 @@
+const portfolio = [
+  {
+    project_id: "1",
+    title: "Project 1",
+    desc: "This project is regarding payments",
+    image: require("./images/palms.jpg")
+  },
+  {
+    project_id: "2",
+    title: "Project 2",
+    desc: "This project is regarding payments",
+    image: require("./images/palms.jpg")
+  },
+  {
+    project_id: "3",
+    title: "Project 3",
+    desc: "This project is regarding payments",
+    image: require("./images/palms.jpg")
+  }
+];
+
+export default portfolio;
